+++ localization/de.lua
@@ -1,0 +1,40 @@
+------------------------------------------------------------
+-- TokenPrice by Sonaza
+-- All rights reserved
+-- http://sonaza.com
+-- German localization provided by Smallinger (https://github.com/Smallinger)
+------------------------------------------------------------
+
+local ADDON_NAME = ...
+local L = LibStub('AceLocale-3.0'):NewLocale(ADDON_NAME, "deDE");
+if not L then return end
+
+L["TokenPrice Options"] = "TokenPreis Optionen";
+L["Show percentage difference in Broker"] = "Prozentwert im Broker anzeigen";
+L["Swap colors (seller/buyer)"] = "Farbe ändern (Verkäufer / Käufer)";
+L["Use literal gold display"] = "Verwende wörtliche Goldanzeige";
+L["Show service prices in gold"] = "Zeige Dienstleistungspreise in Gold an";
+L["Currency Options"] = "Währungs Optionen";
+
+L["TokenPrice"] = "TokenPreis";
+L["Current Price"] = "Aktueller Preis";
+L["Last Change"] = "Letzte Änderung";
+L["%s ago"] = "Vor %s";
+L["Real Price"] = "Echter Preis";
+L["Average Sell Time"] = "Durchschnittliche Verkaufszeit";
+L["Battle.net Balance"] = "Battle.net Guthaben";
+L["Common Service Prices in Gold"] = "Übliche Dienstleistungen in Gold";
+L["Level Boost"] = "Level Boost";
+L["Faction Change"] = "Fraktionswechsel";
+L["Character Transfer"] = "Charaktertransfer";
+L["Race Change"] = "Volkswechsel";
+L["Appearance Change"] = "Charakteranpassung";
+L["Name Change"] = "Charakterumbenennung";
+
+L["Recorded Peaks within 48 hr"] = "Aufgenommene Spitzen innerhalb von 48 Stunden";
+
+L["Highest"] = "Höchste";
+L["Lowest"] = "Unterste";
+L["Note these values are only updated while logged in."] = "Beachte, dass diese Werte nur beim Spielen aktualisiert werden.";
+L["TokenPrice"] = "TokenPreis";
+L["Unavailable"] = "Nicht verfügbar";