+++ localization/en.lua
@@ -1,0 +1,38 @@
+------------------------------------------------------------
+-- TokenPrice by Sonaza
+-- All rights reserved
+-- http://sonaza.com
+------------------------------------------------------------
+
+local ADDON_NAME = ...
+local L = LibStub('AceLocale-3.0'):NewLocale(ADDON_NAME, "enUS", true);
+
+L["TokenPrice Options"] = true;
+L["Show percentage difference in Broker"] = true;
+L["Swap colors (seller/buyer)"] = true;
+L["Use literal gold display"] = true;
+L["Show service prices in gold"] = true;
+L["Currency Options"] = true;
+
+L["TokenPrice"] = true;
+L["Current Price"] = true;
+L["Last Change"] = true;
+L["%s ago"] = true;
+L["Real Price"] = true;
+L["Average Sell Time"] = true;
+L["Battle.net Balance"] = true;
+L["Common Service Prices in Gold"] = true;
+L["Level Boost"] = true;
+L["Faction Change"] = true;
+L["Character Transfer"] = true;
+L["Race Change"] = true;
+L["Appearance Change"] = true;
+L["Name Change"] = true;
+
+L["Recorded Peaks within 48 hr"] = true;
+
+L["Highest"] = true;
+L["Lowest"] = true;
+L["Note these values are only updated while logged in."] = true;
+L["TokenPrice"] = true;
+L["Unavailable"] = true;